+++ src/main/java/com/shootoff/session/io/EventVisitor.java
@@ -10,7 +10,7 @@ public interface EventVisitor {
 	public void visitShot(long timestamp, Shot shot, Optional<Integer> targetIndex, Optional<Integer> hitRegionIndex);
 	public void visitTargetAdd(long timestamp, String targetName);
 	public void visitTargetRemove(long timestamp, int targetIndex);
-	public void visitTargetResize(long timestamp, int targetIndex, int newWidth, int newHeight);
+	public void visitTargetResize(long timestamp, int targetIndex, double newWidth, double newHeight);
 	public void visitTargetMove(long timestamp, int targetIndex, int newX, int newY);
 	public void visitEnd();
 }
